add AppNavigator for UITesting
-refactor code
